iceberg: extract shared manifest utilities from merge_append
Pulls manifest writing, bin-packing, and snapshot construction helpers
out of merge_append_action into reusable utilities, preparing for
row_delta_action which needs the same manifest infrastructure to
commit both data and delete files.
